recheck the existence of std files after the executiong of the job termination cmd
This modification deals with the case:
The error is detected for a running job. Since there is no stdout or stderr file(s) which indicates that the job is already terminated, the program will try to terminate it. But it may happen that the job is somehow terminated due to some reasons while the termination cmd is executed. In this case, termination cmd execution will definitely fail. This previously would lead to __error__ --> __manual__. Here, after the execution of termination cmd, we also re-check the existence of stdout or stderr file(s) such that the above case will be handled properly.

To test this function, __manual__ is created at the current stage. Will be removed in the future.