update __running_job_not_in_queue__
The running job will be tagged with this signal file only after its job id is not found in the job status list consecutively 5 times. This is in case that the system improperly returns a job status list such that many normal running jobs are wrongly judged and tagged with this signal file.