Add a new tag name max_ionic_step for structural relaxation.
max_ionic_step should be -1 (inactive && default) or a possitive integer (active). If max_ionic_step is set, the structural relaxation in which the number ionic steps to reach the ionic convergence is larger than max_ionic_step will be treated as if it does not converge. The corresponding corrections are to change CONTCAR to POSCAR, IBRION=1 and carry out one more round of structural relaxation.